+++ rbx_xml/CHANGELOG.md
@@ -1,10 +1,12 @@
 # rbx_xml Changelog
 
 ## Unreleased
+* Fixed `ProtectedString` support, which was missing from the main deserializer loop
+* Added support for `Content` values
 
 ## 0.3.0 (2019-02-14)
 * Updated `rbx_tree` dependency to `rbx_dom_weak` 0.3.0
-* Added support for ProtectedString, deserializing as String
+* Added support for `ProtectedString`, deserializing as String
 * Added support for `Ref` values ([#8](https://github.com/LPGhatguy/rbx-dom/pull/8))
 
 ## 0.2.0 (2019-01-25)